+++ common-client/src/main/java/de/maxhenkel/voicechat/voice/client/RenderEvents.java
@@ -7,6 +7,7 @@ import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.gui.onboarding.OnboardingManager;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
+import de.maxhenkel.voicechat.plugins.ClientPluginManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.GameRenderer;
@@ -143,6 +144,9 @@ public class RenderEvents {
     }
 
     private void renderPlayerIcon(Player player, Component component, ResourceLocation texture, PoseStack poseStack, MultiBufferSource buffer, int light, float partialTicks) {
+        if (!ClientPluginManager.instance().shouldRenderPlayerIcons(player.getUUID())) {
+            return;
+        }
         Vec3 nametagPos = player.getAttachments().getNullable(EntityAttachment.NAME_TAG, 0, player.getViewYRot(partialTicks));
         if (nametagPos == null) {
             return;