+++ neoforge/changelog.md
@@ -21,3 +21,4 @@
 - Added `isPushToTalkKeyPressed` and `isWhisperKeyPressed` to client API
 - Added `isDisabled` and `isDisconnected` for other players to client API
 - Added `isTalking` and `isWhispering` to client API
+- Added `NameTagIconRenderEvent` to the client API allowing addons to modify or remove the nametag icons